A RDBESDataObject converted from package survey dataset apistrat

This data set is a stratified version of the previous "apiclus2" data. It is derived from the Academic Performance Index computed for all California schools based on standardized testing of students. The original data sets contain information for all schools with at least 100 students and for various probability samples of the data. The design is 1-stage cluster sampling with clusters of unequal sizes. An SRS of 200 districts is selected (psus) from the 755 districts in the population. All schools within district are selected (ssus).

Pckg_survey_apistrat_H1

Format

A list containing entries required for H1 RDBES data:

DE

the Design data table. Contains 1 DE row

SD

the Sampling Details data table. Contains 1 child SD row

VS

the Vessel Selection data table. Contains 200 child rows (the 200 schools finally observed), each associated to its cluster (dname), VSnumberTotalClusters is 755, VSnumberTotal is 50-100 schools sampled

FT

the Fishing Trip data table. Contains 200 child rows (the 200 schools finally observed), each associated to its cluster (dname), FTnumberTotal is the number of schools in the cluster (census)

FO

the Fishing Operation data table. Just 1:1 links to the final data (in SA)

SS

the Species Selection data table. Just 1:1 links to the final data (in SA)

SA

the Sample data table. SAsampleWeightMeasured is enroll

FM

the Frequency Measure data table

BV

the Biological Variable data table

VD

the Vessel Details data table. Contains 311 child rows

SL

the Species List data table. Contains 1 child row

Source

https://CRAN.R-project.org/package=survey