Generic function for estimation of population total and variance

Generic function for estimation of population total and variance

estim(y, enk, enkl, method = "SRSWOR", estFunction, varFunction)

Arguments

y

numeric variable to be estimated

enk

expected value of k

enkl

expected value of k, given l

method

character selection method code e.g SRSWOR

estFunction

the function to use to estimate total given parameters y and enk

varFunction

the function to use to estimate variance given parameters y,enk and enkl

Value

list of 7 elements including the population mean, total (and their variance), the algorithm name used and the I order inclusion probabilities

Examples

estimMC(c(3, 4, 4, 5), c(4, 4, 4, 4), c(8, 8, 8, 8))
#> $est.total
#> [1] 32
#> 
#> $est.mean
#> [1] 4
#> 
#> $est.algorithm
#> [1] "Generalized Horvitz-Thompson aka Mutiple-Count"
#> 
#> $var.total
#> [1] 5.333333
#> 
#> $var.mean
#> [1] 0.08333333
#> 
#> $var.algorithm
#> [1] "Sen-Yates-Grundy"
#> 
#> $PI
#>           [,1]      [,2]      [,3]      [,4]
#> [1,] 0.5000000 0.2142857 0.2142857 0.2142857
#> [2,] 0.2142857 0.5000000 0.2142857 0.2142857
#> [3,] 0.2142857 0.2142857 0.5000000 0.2142857
#> [4,] 0.2142857 0.2142857 0.2142857 0.5000000
#>