Multiple Count Estimator for Population Total and Variance

Multiple Count Estimator for Population Total and Variance

Usage

estimMC(
  y,
  sampled,
  total,
  method = "SRSWOR",
  selProb = NULL,
  incProb = NULL,
  verbose = FALSE
)

Arguments

y

numeric variable to be estimated

sampled

numeric total number of units sampled

total

numeric total number of units int the population

method

character selection method code e.g SRSWOR

selProb

the selection probabilities (if known)

incProb

the inclusion probabilities (if known)

verbose

(Optional) Set to TRUE if you want informative text printed out, or FALSE if you don't. The default is FALSE.

Value

list of 7 elements including the population mean, total (and their variance), the algorithm name used and the I order inclusion probabilities

Examples

estimMC(c(3, 4, 4, 5), c(4, 4, 4, 4), c(8, 8, 8, 8))
#> $est.total
#> [1] 32
#> 
#> $est.mean
#> [1] 4
#> 
#> $est.algorithm
#> [1] "Generalized Horvitz-Thompson aka Mutiple-Count"
#> 
#> $var.total
#> [1] 5.333333
#> 
#> $var.mean
#> [1] 0.08333333
#> 
#> $var.algorithm
#> [1] "Sen-Yates-Grundy"
#> 
#> $PI
#>           [,1]      [,2]      [,3]      [,4]
#> [1,] 0.5000000 0.2142857 0.2142857 0.2142857
#> [2,] 0.2142857 0.5000000 0.2142857 0.2142857
#> [3,] 0.2142857 0.2142857 0.5000000 0.2142857
#> [4,] 0.2142857 0.2142857 0.2142857 0.5000000
#>