Filter an RDBESEstObject

The returned object will include all rows which include the field names and have one of the allowed values in valuesToFilter.

Usage

filterRDBESEstObject(
  RDBESEstObjectToFilter,
  fieldsToFilter,
  valuesToFilter,
  verbose = FALSE
)

Arguments

RDBESEstObjectToFilter

The RDBESEstObject to filter

fieldsToFilter

A vector of the field names you wish to check

valuesToFilter

A vector of the field values you wish to filter for

verbose

(Optional) Set to TRUE if you want informative text printed out, or FALSE if you don't. The default is FALSE.

Value

the filtered input object of the same class as RDBESEstObjectToFilter

Examples

if (FALSE) { # \dontrun{

 myRawObject <- createRDBESDataObject(input = "tests\\testthat\\h1_v_1_19_26")
 myEstObject <- createRDBESEstObject(myRawObject,1)
 myFilteredEst <- filterRDBESEstObject(myEst,c("BVid"),c(7349207))

} # }