Generate a List of Related Data Tables

Generates a named list of data tables that follow the structure of RDBESDataObject. The tables only have columns required for testing The generate tables

generateTestTbls(tblNames, prevTbls = list(), ...)

Arguments

tblNames

character vector of table names to be created

prevTbls

list of data.tables upstream of the generated table. Defaults to empty list

...

Arguments passed on to makeTbl

tblName

Name of the table

rows

numeric number of rows per parent record. Defaults to 4.

propSamp

numeric proportion of how many of total are sampled. This is ignored for "CENSUS". Defaults to 0.5

selMeth

character selection method used. Defaults to "CENSUS". Others like SRSWR or SRSSWOR can be used as well

stratums

character vector of the stratum names to be created. Defaults to c("U"), meaning not stratified.

mean

numeric the expected mean of the target variable. The variable is created using rnorm and saved under column ending with "y". Defaults to 5.

Value

a list of named data.table's

Examples

generateTestTbls(c("A", "B", "C"), selMeth = "SRSWOR")
#> $A
#>    Aid AnumSamp AnumTotal AselectMeth AstratumName Astratification       Ay
#> 1:   1        4         8      SRSWOR            U               N 5.059796
#> 2:   2        4         8      SRSWOR            U               N 5.209115
#> 3:   3        4         8      SRSWOR            U               N 6.091859
#> 4:   4        4         8      SRSWOR            U               N 4.361764
#>    ArecType
#> 1:        A
#> 2:        A
#> 3:        A
#> 4:        A
#> 
#> $B
#>     Bid BnumSamp BnumTotal BselectMeth BstratumName Bstratification Aid
#>  1:   1        4         8      SRSWOR            U               N   1
#>  2:   2        4         8      SRSWOR            U               N   1
#>  3:   3        4         8      SRSWOR            U               N   1
#>  4:   4        4         8      SRSWOR            U               N   1
#>  5:   5        4         8      SRSWOR            U               N   2
#>  6:   6        4         8      SRSWOR            U               N   2
#>  7:   7        4         8      SRSWOR            U               N   2
#>  8:   8        4         8      SRSWOR            U               N   2
#>  9:   9        4         8      SRSWOR            U               N   3
#> 10:  10        4         8      SRSWOR            U               N   3
#> 11:  11        4         8      SRSWOR            U               N   3
#> 12:  12        4         8      SRSWOR            U               N   3
#> 13:  13        4         8      SRSWOR            U               N   4
#> 14:  14        4         8      SRSWOR            U               N   4
#> 15:  15        4         8      SRSWOR            U               N   4
#> 16:  16        4         8      SRSWOR            U               N   4
#>           By BrecType
#>  1: 5.985419        B
#>  2: 5.248652        B
#>  3: 5.039509        B
#>  4: 5.230560        B
#>  5: 3.999534        B
#>  6: 6.509344        B
#>  7: 4.845549        B
#>  8: 4.788430        B
#>  9: 5.927209        B
#> 10: 4.481754        B
#> 11: 3.321447        B
#> 12: 6.738297        B
#> 13: 4.289367        B
#> 14: 3.195323        B
#> 15: 7.582971        B
#> 16: 5.513917        B
#> 
#> $C
#>     Cid CnumSamp CnumTotal CselectMeth CstratumName Cstratification Bid
#>  1:   1        4         8      SRSWOR            U               N   1
#>  2:   2        4         8      SRSWOR            U               N   1
#>  3:   3        4         8      SRSWOR            U               N   1
#>  4:   4        4         8      SRSWOR            U               N   1
#>  5:   5        4         8      SRSWOR            U               N   2
#>  6:   6        4         8      SRSWOR            U               N   2
#>  7:   7        4         8      SRSWOR            U               N   2
#>  8:   8        4         8      SRSWOR            U               N   2
#>  9:   9        4         8      SRSWOR            U               N   3
#> 10:  10        4         8      SRSWOR            U               N   3
#> 11:  11        4         8      SRSWOR            U               N   3
#> 12:  12        4         8      SRSWOR            U               N   3
#> 13:  13        4         8      SRSWOR            U               N   4
#> 14:  14        4         8      SRSWOR            U               N   4
#> 15:  15        4         8      SRSWOR            U               N   4
#> 16:  16        4         8      SRSWOR            U               N   4
#> 17:  17        4         8      SRSWOR            U               N   5
#> 18:  18        4         8      SRSWOR            U               N   5
#> 19:  19        4         8      SRSWOR            U               N   5
#> 20:  20        4         8      SRSWOR            U               N   5
#> 21:  21        4         8      SRSWOR            U               N   6
#> 22:  22        4         8      SRSWOR            U               N   6
#> 23:  23        4         8      SRSWOR            U               N   6
#> 24:  24        4         8      SRSWOR            U               N   6
#> 25:  25        4         8      SRSWOR            U               N   7
#> 26:  26        4         8      SRSWOR            U               N   7
#> 27:  27        4         8      SRSWOR            U               N   7
#> 28:  28        4         8      SRSWOR            U               N   7
#> 29:  29        4         8      SRSWOR            U               N   8
#> 30:  30        4         8      SRSWOR            U               N   8
#> 31:  31        4         8      SRSWOR            U               N   8
#> 32:  32        4         8      SRSWOR            U               N   8
#> 33:  33        4         8      SRSWOR            U               N   9
#> 34:  34        4         8      SRSWOR            U               N   9
#> 35:  35        4         8      SRSWOR            U               N   9
#> 36:  36        4         8      SRSWOR            U               N   9
#> 37:  37        4         8      SRSWOR            U               N  10
#> 38:  38        4         8      SRSWOR            U               N  10
#> 39:  39        4         8      SRSWOR            U               N  10
#> 40:  40        4         8      SRSWOR            U               N  10
#> 41:  41        4         8      SRSWOR            U               N  11
#> 42:  42        4         8      SRSWOR            U               N  11
#> 43:  43        4         8      SRSWOR            U               N  11
#> 44:  44        4         8      SRSWOR            U               N  11
#> 45:  45        4         8      SRSWOR            U               N  12
#> 46:  46        4         8      SRSWOR            U               N  12
#> 47:  47        4         8      SRSWOR            U               N  12
#> 48:  48        4         8      SRSWOR            U               N  12
#> 49:  49        4         8      SRSWOR            U               N  13
#> 50:  50        4         8      SRSWOR            U               N  13
#> 51:  51        4         8      SRSWOR            U               N  13
#> 52:  52        4         8      SRSWOR            U               N  13
#> 53:  53        4         8      SRSWOR            U               N  14
#> 54:  54        4         8      SRSWOR            U               N  14
#> 55:  55        4         8      SRSWOR            U               N  14
#> 56:  56        4         8      SRSWOR            U               N  14
#> 57:  57        4         8      SRSWOR            U               N  15
#> 58:  58        4         8      SRSWOR            U               N  15
#> 59:  59        4         8      SRSWOR            U               N  15
#> 60:  60        4         8      SRSWOR            U               N  15
#> 61:  61        4         8      SRSWOR            U               N  16
#> 62:  62        4         8      SRSWOR            U               N  16
#> 63:  63        4         8      SRSWOR            U               N  16
#> 64:  64        4         8      SRSWOR            U               N  16
#>     Cid CnumSamp CnumTotal CselectMeth CstratumName Cstratification Bid
#>           Cy CrecType
#>  1: 4.484684        C
#>  2: 5.705811        C
#>  3: 5.899654        C
#>  4: 4.693181        C
#>  5: 4.688853        C
#>  6: 3.937041        C
#>  7: 6.527700        C
#>  8: 6.094963        C
#>  9: 3.338404        C
#> 10: 3.987265        C
#> 11: 5.851209        C
#> 12: 5.989754        C
#> 13: 6.215717        C
#> 14: 4.038703        C
#> 15: 4.636711        C
#> 16: 4.565166        C
#> 17: 5.204587        C
#> 18: 6.865241        C
#> 19: 4.958473        C
#> 20: 3.834345        C
#> 21: 5.800557        C
#> 22: 5.400122        C
#> 23: 4.249710        C
#> 24: 4.946107        C
#> 25: 5.073917        C
#> 26: 5.424751        C
#> 27: 4.186104        C
#> 28: 3.933887        C
#> 29: 5.052083        C
#> 30: 4.513774        C
#> 31: 4.680176        C
#> 32: 3.765845        C
#> 33: 4.244430        C
#> 34: 2.968827        C
#> 35: 4.645050        C
#> 36: 4.151347        C
#> 37: 3.748685        C
#> 38: 5.143390        C
#> 39: 4.883471        C
#> 40: 4.677257        C
#> 41: 6.194899        C
#> 42: 5.189988        C
#> 43: 3.702760        C
#> 44: 5.614124        C
#> 45: 5.930687        C
#> 46: 4.569738        C
#> 47: 7.156580        C
#> 48: 5.721332        C
#> 49: 4.782293        C
#> 50: 4.091273        C
#> 51: 4.648011        C
#> 52: 3.009739        C
#> 53: 5.711641        C
#> 54: 5.013260        C
#> 55: 5.071630        C
#> 56: 4.802016        C
#> 57: 4.919502        C
#> 58: 6.331971        C
#> 59: 4.132300        C
#> 60: 3.914042        C
#> 61: 5.411644        C
#> 62: 6.428249        C
#> 63: 5.063158        C
#> 64: 4.881505        C
#>           Cy CrecType
#> 
generateTestTbls(LETTERS[1:5]) # makes 5 tables with method CENSUS
#> $A
#>    Aid AnumSamp AnumTotal AselectMeth AstratumName Astratification       Ay
#> 1:   1        4         4      CENSUS            U               N 6.161942
#> 2:   2        4         4      CENSUS            U               N 5.819739
#> 3:   3        4         4      CENSUS            U               N 4.531657
#> 4:   4        4         4      CENSUS            U               N 4.655372
#>    ArecType
#> 1:        A
#> 2:        A
#> 3:        A
#> 4:        A
#> 
#> $B
#>     Bid BnumSamp BnumTotal BselectMeth BstratumName Bstratification Aid
#>  1:   1        4         4      CENSUS            U               N   1
#>  2:   2        4         4      CENSUS            U               N   1
#>  3:   3        4         4      CENSUS            U               N   1
#>  4:   4        4         4      CENSUS            U               N   1
#>  5:   5        4         4      CENSUS            U               N   2
#>  6:   6        4         4      CENSUS            U               N   2
#>  7:   7        4         4      CENSUS            U               N   2
#>  8:   8        4         4      CENSUS            U               N   2
#>  9:   9        4         4      CENSUS            U               N   3
#> 10:  10        4         4      CENSUS            U               N   3
#> 11:  11        4         4      CENSUS            U               N   3
#> 12:  12        4         4      CENSUS            U               N   3
#> 13:  13        4         4      CENSUS            U               N   4
#> 14:  14        4         4      CENSUS            U               N   4
#> 15:  15        4         4      CENSUS            U               N   4
#> 16:  16        4         4      CENSUS            U               N   4
#>           By BrecType
#>  1: 3.159089        B
#>  2: 4.204945        B
#>  3: 5.892407        B
#>  4: 5.182575        B
#>  5: 5.651843        B
#>  6: 6.309282        B
#>  7: 6.968059        B
#>  8: 4.489842        B
#>  9: 4.262068        B
#> 10: 5.081843        B
#> 11: 3.962412        B
#> 12: 4.895395        B
#> 13: 5.171659        B
#> 14: 4.621888        B
#> 15: 5.054465        B
#> 16: 5.009659        B
#> 
#> $C
#>     Cid CnumSamp CnumTotal CselectMeth CstratumName Cstratification Bid
#>  1:   1        4         4      CENSUS            U               N   1
#>  2:   2        4         4      CENSUS            U               N   1
#>  3:   3        4         4      CENSUS            U               N   1
#>  4:   4        4         4      CENSUS            U               N   1
#>  5:   5        4         4      CENSUS            U               N   2
#>  6:   6        4         4      CENSUS            U               N   2
#>  7:   7        4         4      CENSUS            U               N   2
#>  8:   8        4         4      CENSUS            U               N   2
#>  9:   9        4         4      CENSUS            U               N   3
#> 10:  10        4         4      CENSUS            U               N   3
#> 11:  11        4         4      CENSUS            U               N   3
#> 12:  12        4         4      CENSUS            U               N   3
#> 13:  13        4         4      CENSUS            U               N   4
#> 14:  14        4         4      CENSUS            U               N   4
#> 15:  15        4         4      CENSUS            U               N   4
#> 16:  16        4         4      CENSUS            U               N   4
#> 17:  17        4         4      CENSUS            U               N   5
#> 18:  18        4         4      CENSUS            U               N   5
#> 19:  19        4         4      CENSUS            U               N   5
#> 20:  20        4         4      CENSUS            U               N   5
#> 21:  21        4         4      CENSUS            U               N   6
#> 22:  22        4         4      CENSUS            U               N   6
#> 23:  23        4         4      CENSUS            U               N   6
#> 24:  24        4         4      CENSUS            U               N   6
#> 25:  25        4         4      CENSUS            U               N   7
#> 26:  26        4         4      CENSUS            U               N   7
#> 27:  27        4         4      CENSUS            U               N   7
#> 28:  28        4         4      CENSUS            U               N   7
#> 29:  29        4         4      CENSUS            U               N   8
#> 30:  30        4         4      CENSUS            U               N   8
#> 31:  31        4         4      CENSUS            U               N   8
#> 32:  32        4         4      CENSUS            U               N   8
#> 33:  33        4         4      CENSUS            U               N   9
#> 34:  34        4         4      CENSUS            U               N   9
#> 35:  35        4         4      CENSUS            U               N   9
#> 36:  36        4         4      CENSUS            U               N   9
#> 37:  37        4         4      CENSUS            U               N  10
#> 38:  38        4         4      CENSUS            U               N  10
#> 39:  39        4         4      CENSUS            U               N  10
#> 40:  40        4         4      CENSUS            U               N  10
#> 41:  41        4         4      CENSUS            U               N  11
#> 42:  42        4         4      CENSUS            U               N  11
#> 43:  43        4         4      CENSUS            U               N  11
#> 44:  44        4         4      CENSUS            U               N  11
#> 45:  45        4         4      CENSUS            U               N  12
#> 46:  46        4         4      CENSUS            U               N  12
#> 47:  47        4         4      CENSUS            U               N  12
#> 48:  48        4         4      CENSUS            U               N  12
#> 49:  49        4         4      CENSUS            U               N  13
#> 50:  50        4         4      CENSUS            U               N  13
#> 51:  51        4         4      CENSUS            U               N  13
#> 52:  52        4         4      CENSUS            U               N  13
#> 53:  53        4         4      CENSUS            U               N  14
#> 54:  54        4         4      CENSUS            U               N  14
#> 55:  55        4         4      CENSUS            U               N  14
#> 56:  56        4         4      CENSUS            U               N  14
#> 57:  57        4         4      CENSUS            U               N  15
#> 58:  58        4         4      CENSUS            U               N  15
#> 59:  59        4         4      CENSUS            U               N  15
#> 60:  60        4         4      CENSUS            U               N  15
#> 61:  61        4         4      CENSUS            U               N  16
#> 62:  62        4         4      CENSUS            U               N  16
#> 63:  63        4         4      CENSUS            U               N  16
#> 64:  64        4         4      CENSUS            U               N  16
#>     Cid CnumSamp CnumTotal CselectMeth CstratumName Cstratification Bid
#>           Cy CrecType
#>  1: 4.158755        C
#>  2: 3.931504        C
#>  3: 3.670558        C
#>  4: 5.121128        C
#>  5: 6.661482        C
#>  6: 6.218617        C
#>  7: 5.444234        C
#>  8: 4.933944        C
#>  9: 3.927189        C
#> 10: 3.542432        C
#> 11: 3.802181        C
#> 12: 6.066261        C
#> 13: 4.942256        C
#> 14: 5.739374        C
#> 15: 3.154376        C
#> 16: 3.740846        C
#> 17: 5.349956        C
#> 18: 5.799147        C
#> 19: 5.644117        C
#> 20: 4.985471        C
#> 21: 5.024714        C
#> 22: 5.150868        C
#> 23: 4.561464        C
#> 24: 4.897452        C
#> 25: 4.562911        C
#> 26: 4.333080        C
#> 27: 5.535671        C
#> 28: 5.577274        C
#> 29: 5.460661        C
#> 30: 3.753247        C
#> 31: 6.239514        C
#> 32: 6.551325        C
#> 33: 4.338883        C
#> 34: 5.165887        C
#> 35: 4.441775        C
#> 36: 6.177045        C
#> 37: 4.088513        C
#> 38: 5.299325        C
#> 39: 4.599932        C
#> 40: 2.213818        C
#> 41: 5.090738        C
#> 42: 5.292740        C
#> 43: 5.945867        C
#> 44: 4.272896        C
#> 45: 5.462731        C
#> 46: 4.907170        C
#> 47: 5.343069        C
#> 48: 4.972722        C
#> 49: 4.364677        C
#> 50: 4.941094        C
#> 51: 3.322401        C
#> 52: 4.347475        C
#> 53: 2.227744        C
#> 54: 4.544975        C
#> 55: 6.966690        C
#> 56: 4.008519        C
#> 57: 3.858874        C
#> 58: 4.106771        C
#> 59: 5.566173        C
#> 60: 5.290341        C
#> 61: 6.380154        C
#> 62: 4.726902        C
#> 63: 5.427248        C
#> 64: 6.629413        C
#>           Cy CrecType
#> 
#> $D
#>      Did DnumSamp DnumTotal DselectMeth DstratumName Dstratification Cid
#>   1:   1        4         4      CENSUS            U               N   1
#>   2:   2        4         4      CENSUS            U               N   1
#>   3:   3        4         4      CENSUS            U               N   1
#>   4:   4        4         4      CENSUS            U               N   1
#>   5:   5        4         4      CENSUS            U               N   2
#>  ---                                                                    
#> 252: 252        4         4      CENSUS            U               N  63
#> 253: 253        4         4      CENSUS            U               N  64
#> 254: 254        4         4      CENSUS            U               N  64
#> 255: 255        4         4      CENSUS            U               N  64
#> 256: 256        4         4      CENSUS            U               N  64
#>            Dy DrecType
#>   1: 3.416239        D
#>   2: 4.146988        D
#>   3: 2.977335        D
#>   4: 4.342866        D
#>   5: 2.990748        D
#>  ---                  
#> 252: 6.953975        D
#> 253: 6.570608        D
#> 254: 4.497441        D
#> 255: 3.901114        D
#> 256: 5.322199        D
#> 
#> $E
#>        Eid EnumSamp EnumTotal EselectMeth EstratumName Estratification Did
#>    1:    1        4         4      CENSUS            U               N   1
#>    2:    2        4         4      CENSUS            U               N   1
#>    3:    3        4         4      CENSUS            U               N   1
#>    4:    4        4         4      CENSUS            U               N   1
#>    5:    5        4         4      CENSUS            U               N   2
#>   ---                                                                     
#> 1020: 1020        4         4      CENSUS            U               N 255
#> 1021: 1021        4         4      CENSUS            U               N 256
#> 1022: 1022        4         4      CENSUS            U               N 256
#> 1023: 1023        4         4      CENSUS            U               N 256
#> 1024: 1024        4         4      CENSUS            U               N 256
#>             Ey ErecType
#>    1: 5.645929        E
#>    2: 7.405617        E
#>    3: 4.698105        E
#>    4: 5.346046        E
#>    5: 3.856200        E
#>   ---                  
#> 1020: 5.389098        E
#> 1021: 4.915865        E
#> 1022: 3.371979        E
#> 1023: 4.600518        E
#> 1024: 3.117376        E
#>