Convert List of Data Frames to a RDBES Data Object

This function converts a list of data frames into an object of class RDBESDataObject.

importRDBESDataDFS(myList, castToCorrectDataTypes = TRUE, strict = TRUE, ...)

Arguments

myList

A list of data tables. Each element of the list should be a data frame with an RDBES two-letter name (e.g. "DE").

castToCorrectDataTypes

logical. Indicates whether to cast the columns to the correct data types. Default is TRUE.

strict

logical. Indicates level of validation of the RDBESDataObject it creates - should the validation be strict? Default is TRUE.

Value

An RDBESDataObject with each element being a data table.

Details

Tables in the input list should have the correct 2-letter RDBES name (e.g. "DE", "LE", etc.). The function converts all data frames to data.table. NULL tables are left as NULL.

If castToCorrectDataTypes = TRUE, it ensures all columns are of the correct data type using setRDBESDataObjectDataTypes.

Column names are replaced with the RDBES 'R names' from the model documentation.

The function then sets a key on each table using the 'XXid' column as the key, where 'XX' is the name of that table. It alos replaces all empty strings with NA.

It then uses the newRDBESDataObject function to create a new RDBESDataObject from the input.

Finally, it validates the RDBESDataObject using RDBEScore::validateRDBESDataObject and returns it.