Get lower table data from upper table id

This function takes an upper table field name, values for that field, a list of tables, and a target lower level table name. It returns the rows of the target lower level table that are associated with the given values of the upper table field. NB! running on the RDBESDataObject will work properly if it is sorted by default the tables are not in the correct order for a specific hierarchy.

Usage

lowerTblData(field, values, tbls, level, verbose = FALSE, path_order = NULL)

Arguments

field

A character string specifying the name of the upper table field.

values

A vector of values for the upper table field.

tbls

A named list of data frames representing the tables.

level

A character string specifying the name of the target lower level table.

verbose

A logical value indicating whether to print intermediate levels during recursion.

Value

A data frame containing the rows of the target lower level table that are associated with the given values of the upper table field.

Examples

# it is important to run the function only on the sorted RDBESDataObject
RDBEScore:::lowerTblData("TEid", c(4), sort(H8ExampleEE1), "SA", TRUE)
#> TEid: 4
#> VSid: 5, 6
#> LEid: 5, 6
#> SSid: 5, 6
#>     TEid  VSid  LEid  SSid  SAid SAparSequNum SArecType SAseqNum
#>    <num> <num> <num> <num> <num>       <lgcl>    <char>    <num>
#> 1:     4     5     5     5     5           NA        SA        5
#> 2:     4     6     6     6     6           NA        SA        6
#>    SAstratification SAstratumName SAspeCode SAspeCodeFAO SApres SAstateOfProc
#>              <char>        <char>     <num>       <char> <char>        <char>
#> 1:                N             U    126425          SPR    WHL           FRE
#> 2:                N             U    126425          SPR    WHL           FRE
#>    SAcatchCat SAlandCat SAcommCatScl SAcommCat  SAsex SAunitName SAunitType
#>        <char>    <char>       <lgcl>    <lgcl> <char>     <char>     <char>
#> 1:        Lan       HuC           NA        NA      U     202205        Box
#> 2:        Lan       HuC           NA        NA      U     202206        Box
#>    SAtotalWtLive SAsampWtLive SAnumTotal SAnumSamp SAselProb SAselectMeth
#>           <lgcl>       <lgcl>      <num>     <int>    <lgcl>       <char>
#> 1:            NA           NA        794         1        NA       SRSWOR
#> 2:            NA           NA       2805         1        NA       SRSWOR
#>    SAlowHierarchy    SAsampler SAnoSampReasonFM SAnoSampReasonBV SAtotalWtMes
#>            <char>       <char>           <lgcl>           <lgcl>        <num>
#> 1:              C SelfSampling               NA               NA       649000
#> 2:              C SelfSampling               NA               NA      1170000
#>    SAtotWtMeaDatBas SAsampWtMes            SAspecState  SAgsaSubarea SAarea
#>              <char>       <num>                 <char>        <char> <lgcl>
#> 1:         Official         817 DeadOrZeroProbSurvival NotApplicable     NA
#> 2:         Official         417 DeadOrZeroProbSurvival NotApplicable     NA
#>    SAstatRect SAeconZoneIndi SAjurisdArea SAgeoDatBas SAgeoSou SAnatFishAct
#>        <lgcl>         <lgcl>       <lgcl>      <char>   <char>       <lgcl>
#> 1:         NA             NA           NA        <NA>     <NA>           NA
#> 2:         NA             NA           NA        <NA>     <NA>           NA
#>    SAmetier5 SAmetier6 SAgear SAgeaDatBas SAgearSou SAmeshSize SAselDev
#>       <lgcl>    <lgcl> <lgcl>      <char>    <char>     <lgcl>   <lgcl>
#> 1:        NA        NA     NA        <NA>      <NA>         NA       NA
#> 2:        NA        NA     NA        <NA>      <NA>         NA       NA
#>    SAselDevMeshSize SAincProb SAsamp SAconFacMeasLive SAnoSampReason
#>              <lgcl>    <lgcl> <char>           <lgcl>         <lgcl>
#> 1:               NA        NA      Y               NA             NA
#> 2:               NA        NA      Y               NA             NA
#>    SAnonRespCol SAauxVarTot SAauxVarValue SAauxVarName SAauxVarUnit
#>          <char>      <lgcl>        <lgcl>       <lgcl>       <lgcl>
#> 1:            N          NA            NA           NA           NA
#> 2:            N          NA            NA           NA           NA
#>    SAfishManUnit
#>           <lgcl>
#> 1:            NA
#> 2:            NA

DE <- data.table::data.table(DEid = c(1, 2, 3, 4), SDid = c(1, 2, 3, 4))
SD <- data.table::data.table(SDid = c(1, 2, 3, 4), TEid = c(1, 2, 3, 4))
TE <- data.table::data.table(SDid = c(1, 2, 3, 4), TEid = c(1, 2, 3, 4))
VS <- data.table::data.table(TEid = c(1, 2, 3, 4), VSid = c(1, 2, 3, 4))
LE <- data.table::data.table(VSid = 1:5, LEid = 1:5, value = c(10, 20, 3, 4, 6))
tblsSprat <- list( DE = DE ,SD = SD, TE = TE, VS = VS, LE = LE )

RDBEScore:::lowerTblData("TEid", c(4), tblsSprat, "LE", TRUE)
#> TEid: 4
#> VSid: 4
#>     TEid  VSid  LEid value
#>    <num> <int> <int> <num>
#> 1:     4     4     4     4