Print method for RDBESDataObject

This method prints the hierarchy of the DE data.table (if it exists), and the number of rows for each data.table in the RDBESDataObject that is not NULL. It also provides the sampling method and number sampled and number total for tables where it is applicable. If the RDBESDataObject has a mixed hierarchy, a warning message is printed.

This method sorts the RDBESDataObject based on the hierarchy.

This method returns a list containing the hierarchy of the DE data.table, the number of rows for each data.table in the RDBESDataObject that is not NULL, and a logical value indicating if the hierarchy is not NULL.

Usage

# S3 method for class 'RDBESDataObject'
print(x, ...)

# S3 method for class 'RDBESDataObject'
sort(x, decreasing = TRUE, ...)

# S3 method for class 'RDBESDataObject'
summary(object, ...)

Arguments

x

An object of class RDBESDataObject.

...

parameters to underling functions (not used currently)

decreasing

should hierarchy tables be the first ones

object

An object of class RDBESDataObject.

Value

None.

The sorted RDBESDataObject by hierarchy.

A list with three elements:

hierarchy: The hierarchy of the DE data.table in the RDBESDataObject.

rows: A named list where the names are the names of the data.tables in the RDBESDataObject and the values are the number of rows in each data.table. NULL values are excluded.

CS: A logical value indicating if the hierarchy is not NULL.

Examples

# Print the package data object
print(H1Example)
#> Hierarchy 1 RDBESdataObject:
#>  DE: 8
#>  SD: 8
#>  VS: 1214 (SRSWOR,CENSUS,SRSWR: 2-135/4-1382)
#>  FT: 1430 (CENSUS,SRSWR: 1-3/1-100)
#>  FO: 1916 (CENSUS,SRSWR: 1-3/1-20)
#>  SS: 1916 (CENSUS,SRSWR: 1/1-4)
#>  SA: 1916 (CENSUS,SRSWR: 1/1-2)
#>  FM: 7290
#>  BV: 14580 (SRSWR: 2/2)
#>  VD: 311
#>  SL: 6
#>  IS: 25
# Sort the package data
sort(H8ExampleEE1)
#> Hierarchy 8 RDBESdataObject:
#>  DE: 1
#>  SD: 1
#>  TE: 11 (SRSWOR: 2-3/4)
#>  VS: 15 (SRSWR: 1-2/7-12)
#>  LE: 15 (SRSWOR: 1/1)
#>  SS: 15 (CENSUS: 1/1)
#>  SA: 15 (SRSWOR: 1/794-14268)
#>  BV: 3995 (CENSUS: 16-100/16-100)
#>  VD: 7
#>  SL: 1
#>  IS: 2
#>  CL: 71
#>  CE: 132
# Get summary of the package data
summary(H1Example)
#> $hierarchy
#> [1] 1
#> 
#> $data
#> $data$DE
#> $data$DE$design
#>   year
#> 1 1965
#> 
#> $data$DE$rows
#> [1] 8
#> 
#> 
#> $data$SD
#> $data$SD$design
#> data frame with 0 columns and 0 rows
#> 
#> $data$SD$rows
#> [1] 8
#> 
#> 
#> $data$VS
#> $data$VS$design
#>    numTotal numSamp selectMeth
#> 1      1054     103     SRSWOR
#> 2       220      21     SRSWOR
#> 3      1382     135     SRSWOR
#> 4       422      41     SRSWOR
#> 5         4       4     CENSUS
#> 6        13       2     SRSWOR
#> 7        10       2     SRSWOR
#> 8         8       2     SRSWOR
#> 9        11       2     SRSWOR
#> 10        9       2     SRSWOR
#> 11        6       2     SRSWOR
#> 12        7       2     SRSWOR
#> 13        5       2     SRSWOR
#> 14       12       2     SRSWOR
#> 15      163      20     SRSWOR
#> 16      109      20     SRSWOR
#> 17      219      20     SRSWOR
#> 18      318      20     SRSWOR
#> 19      259      20     SRSWOR
#> 20      311      20     SRSWOR
#> 21      263      20     SRSWOR
#> 22      180      20     SRSWOR
#> 23      114      20     SRSWOR
#> 24      367      20     SRSWOR
#> 25       20      20     CENSUS
#> 26       34      34     CENSUS
#> 27       26      26     CENSUS
#> 28       28      28     CENSUS
#> 29       19      19     CENSUS
#> 30       32      32     CENSUS
#> 31       17      17     CENSUS
#> 32       21      21     CENSUS
#> 33       24      24     CENSUS
#> 34       60       3      SRSWR
#> 
#> $data$VS$rows
#> [1] 1214
#> 
#> 
#> $data$FT
#> $data$FT$design
#>   numTotal numSamp selectMeth
#> 1        1       1     CENSUS
#> 2      100       3      SRSWR
#> 
#> $data$FT$rows
#> [1] 1430
#> 
#> 
#> $data$FO
#> $data$FO$design
#>   numTotal numSamp selectMeth
#> 1        1       1     CENSUS
#> 2       20       3      SRSWR
#> 
#> $data$FO$rows
#> [1] 1916
#> 
#> 
#> $data$SS
#> $data$SS$design
#>   numTotal numSamp selectMeth
#> 1        1       1     CENSUS
#> 2        4       1      SRSWR
#> 
#> $data$SS$rows
#> [1] 1916
#> 
#> 
#> $data$SA
#> $data$SA$design
#>   numTotal numSamp selectMeth
#> 1        1       1     CENSUS
#> 2        2       1      SRSWR
#> 
#> $data$SA$rows
#> [1] 1916
#> 
#> 
#> $data$FM
#> $data$FM$design
#> data frame with 0 columns and 0 rows
#> 
#> $data$FM$rows
#> [1] 7290
#> 
#> 
#> $data$BV
#> $data$BV$design
#>   numTotal numSamp selectMeth
#> 1        2       2      SRSWR
#> 
#> $data$BV$rows
#> [1] 14580
#> 
#> 
#> $data$VD
#> $data$VD$design
#>   year
#> 1 1965
#> 
#> $data$VD$rows
#> [1] 311
#> 
#> 
#> $data$SL
#> $data$SL$design
#>   year
#> 1 1965
#> 
#> $data$SL$rows
#> [1] 6
#> 
#> 
#> $data$IS
#> $data$IS$design
#> data frame with 0 columns and 0 rows
#> 
#> $data$IS$rows
#> [1] 25
#> 
#> 
#>