Generate a List of Related Data Tables

Generates a named list of data tables that follow the structure of RDBESDataObject. The tables only have columns required for testing The generate tables

Usage

generateTestTbls(tblNames, prevTbls = list(), ...)

Arguments

tblNames

character vector of table names to be created

prevTbls

list of data.tables upstream of the generated table. Defaults to empty list

...

Arguments passed on to makeTbl

tblName

Name of the table

rows

numeric number of rows per parent record. Defaults to 4.

propSamp

numeric proportion of how many of total are sampled. This is ignored for "CENSUS". Defaults to 0.5

selMeth

character selection method used. Defaults to "CENSUS". Others like SRSWR or SRSSWOR can be used as well

stratums

character vector of the stratum names to be created. Defaults to c("U"), meaning not stratified.

mean

numeric the expected mean of the target variable. The variable is created using rnorm and saved under column ending with "y". Defaults to 5.

Value

a list of named data.table's

Examples

if (FALSE) { # \dontrun{
generateTestTbls(c("A", "B", "C"), selMeth = "SRSWOR")
generateTestTbls(LETTERS[1:5]) # makes 5 tables with method CENSUS
} # }