Generate a List of Related Data Tables

Generates a named list of data tables that follow the structure of RDBESDataObject. The tables only have columns required for testing The generate tables

generateTestTbls(tblNames, prevTbls = list(), ...)

Arguments

tblNames

character vector of table names to be created

prevTbls

list of data.tables upstream of the generated table. Defaults to empty list

...

Arguments passed on to makeTbl

tblName

Name of the table

rows

numeric number of rows per parent record. Defaults to 4.

propSamp

numeric proportion of how many of total are sampled. This is ignored for "CENSUS". Defaults to 0.5

selMeth

character selection method used. Defaults to "CENSUS". Others like SRSWR or SRSSWOR can be used as well

stratums

character vector of the stratum names to be created. Defaults to c("U"), meaning not stratified.

mean

numeric the expected mean of the target variable. The variable is created using rnorm and saved under column ending with "y". Defaults to 5.

Value

a list of named data.table's

Examples

generateTestTbls(c("A", "B", "C"), selMeth = "SRSWOR")
#> $A
#>    Aid AnumSamp AnumTotal AselectMeth AstratumName Astratification       Ay
#> 1:   1        4         8      SRSWOR            U               N 5.944787
#> 2:   2        4         8      SRSWOR            U               N 5.371273
#> 3:   3        4         8      SRSWOR            U               N 5.440532
#> 4:   4        4         8      SRSWOR            U               N 6.407300
#>    ArecType
#> 1:        A
#> 2:        A
#> 3:        A
#> 4:        A
#> 
#> $B
#>     Bid BnumSamp BnumTotal BselectMeth BstratumName Bstratification Aid
#>  1:   1        4         8      SRSWOR            U               N   1
#>  2:   2        4         8      SRSWOR            U               N   1
#>  3:   3        4         8      SRSWOR            U               N   1
#>  4:   4        4         8      SRSWOR            U               N   1
#>  5:   5        4         8      SRSWOR            U               N   2
#>  6:   6        4         8      SRSWOR            U               N   2
#>  7:   7        4         8      SRSWOR            U               N   2
#>  8:   8        4         8      SRSWOR            U               N   2
#>  9:   9        4         8      SRSWOR            U               N   3
#> 10:  10        4         8      SRSWOR            U               N   3
#> 11:  11        4         8      SRSWOR            U               N   3
#> 12:  12        4         8      SRSWOR            U               N   3
#> 13:  13        4         8      SRSWOR            U               N   4
#> 14:  14        4         8      SRSWOR            U               N   4
#> 15:  15        4         8      SRSWOR            U               N   4
#> 16:  16        4         8      SRSWOR            U               N   4
#>           By BrecType
#>  1: 4.656553        B
#>  2: 5.170356        B
#>  3: 4.698289        B
#>  4: 4.300422        B
#>  5: 4.163244        B
#>  6: 3.670099        B
#>  7: 5.203176        B
#>  8: 3.753104        B
#>  9: 4.542708        B
#> 10: 5.622110        B
#> 11: 5.325732        B
#> 12: 4.871415        B
#> 13: 4.003677        B
#> 14: 6.005111        B
#> 15: 4.327454        B
#> 16: 4.421267        B
#> 
#> $C
#>     Cid CnumSamp CnumTotal CselectMeth CstratumName Cstratification Bid
#>  1:   1        4         8      SRSWOR            U               N   1
#>  2:   2        4         8      SRSWOR            U               N   1
#>  3:   3        4         8      SRSWOR            U               N   1
#>  4:   4        4         8      SRSWOR            U               N   1
#>  5:   5        4         8      SRSWOR            U               N   2
#>  6:   6        4         8      SRSWOR            U               N   2
#>  7:   7        4         8      SRSWOR            U               N   2
#>  8:   8        4         8      SRSWOR            U               N   2
#>  9:   9        4         8      SRSWOR            U               N   3
#> 10:  10        4         8      SRSWOR            U               N   3
#> 11:  11        4         8      SRSWOR            U               N   3
#> 12:  12        4         8      SRSWOR            U               N   3
#> 13:  13        4         8      SRSWOR            U               N   4
#> 14:  14        4         8      SRSWOR            U               N   4
#> 15:  15        4         8      SRSWOR            U               N   4
#> 16:  16        4         8      SRSWOR            U               N   4
#> 17:  17        4         8      SRSWOR            U               N   5
#> 18:  18        4         8      SRSWOR            U               N   5
#> 19:  19        4         8      SRSWOR            U               N   5
#> 20:  20        4         8      SRSWOR            U               N   5
#> 21:  21        4         8      SRSWOR            U               N   6
#> 22:  22        4         8      SRSWOR            U               N   6
#> 23:  23        4         8      SRSWOR            U               N   6
#> 24:  24        4         8      SRSWOR            U               N   6
#> 25:  25        4         8      SRSWOR            U               N   7
#> 26:  26        4         8      SRSWOR            U               N   7
#> 27:  27        4         8      SRSWOR            U               N   7
#> 28:  28        4         8      SRSWOR            U               N   7
#> 29:  29        4         8      SRSWOR            U               N   8
#> 30:  30        4         8      SRSWOR            U               N   8
#> 31:  31        4         8      SRSWOR            U               N   8
#> 32:  32        4         8      SRSWOR            U               N   8
#> 33:  33        4         8      SRSWOR            U               N   9
#> 34:  34        4         8      SRSWOR            U               N   9
#> 35:  35        4         8      SRSWOR            U               N   9
#> 36:  36        4         8      SRSWOR            U               N   9
#> 37:  37        4         8      SRSWOR            U               N  10
#> 38:  38        4         8      SRSWOR            U               N  10
#> 39:  39        4         8      SRSWOR            U               N  10
#> 40:  40        4         8      SRSWOR            U               N  10
#> 41:  41        4         8      SRSWOR            U               N  11
#> 42:  42        4         8      SRSWOR            U               N  11
#> 43:  43        4         8      SRSWOR            U               N  11
#> 44:  44        4         8      SRSWOR            U               N  11
#> 45:  45        4         8      SRSWOR            U               N  12
#> 46:  46        4         8      SRSWOR            U               N  12
#> 47:  47        4         8      SRSWOR            U               N  12
#> 48:  48        4         8      SRSWOR            U               N  12
#> 49:  49        4         8      SRSWOR            U               N  13
#> 50:  50        4         8      SRSWOR            U               N  13
#> 51:  51        4         8      SRSWOR            U               N  13
#> 52:  52        4         8      SRSWOR            U               N  13
#> 53:  53        4         8      SRSWOR            U               N  14
#> 54:  54        4         8      SRSWOR            U               N  14
#> 55:  55        4         8      SRSWOR            U               N  14
#> 56:  56        4         8      SRSWOR            U               N  14
#> 57:  57        4         8      SRSWOR            U               N  15
#> 58:  58        4         8      SRSWOR            U               N  15
#> 59:  59        4         8      SRSWOR            U               N  15
#> 60:  60        4         8      SRSWOR            U               N  15
#> 61:  61        4         8      SRSWOR            U               N  16
#> 62:  62        4         8      SRSWOR            U               N  16
#> 63:  63        4         8      SRSWOR            U               N  16
#> 64:  64        4         8      SRSWOR            U               N  16
#>     Cid CnumSamp CnumTotal CselectMeth CstratumName Cstratification Bid
#>           Cy CrecType
#>  1: 5.888166        C
#>  2: 5.396265        C
#>  3: 4.846081        C
#>  4: 5.516464        C
#>  5: 5.355454        C
#>  6: 6.458055        C
#>  7: 6.705419        C
#>  8: 4.655582        C
#>  9: 5.147100        C
#> 10: 5.213869        C
#> 11: 6.555318        C
#> 12: 4.552498        C
#> 13: 5.733811        C
#> 14: 4.994587        C
#> 15: 5.579633        C
#> 16: 4.056036        C
#> 17: 5.527346        C
#> 18: 4.708313        C
#> 19: 5.000436        C
#> 20: 3.680369        C
#> 21: 5.758115        C
#> 22: 6.397123        C
#> 23: 4.438231        C
#> 24: 5.498436        C
#> 25: 4.089672        C
#> 26: 3.638708        C
#> 27: 6.105329        C
#> 28: 6.154216        C
#> 29: 5.014171        C
#> 30: 3.120042        C
#> 31: 3.239949        C
#> 32: 5.945615        C
#> 33: 4.266604        C
#> 34: 4.496304        C
#> 35: 3.610381        C
#> 36: 1.717506        C
#> 37: 5.846342        C
#> 38: 4.737916        C
#> 39: 4.364669        C
#> 40: 5.250314        C
#> 41: 4.557128        C
#> 42: 6.329709        C
#> 43: 5.502553        C
#> 44: 3.885474        C
#> 45: 4.498330        C
#> 46: 4.711079        C
#> 47: 5.440136        C
#> 48: 5.857625        C
#> 49: 5.832026        C
#> 50: 5.221951        C
#> 51: 6.976548        C
#> 52: 4.784210        C
#> 53: 7.450113        C
#> 54: 4.707623        C
#> 55: 4.199770        C
#> 56: 6.614522        C
#> 57: 5.955169        C
#> 58: 4.929773        C
#> 59: 5.343942        C
#> 60: 5.212628        C
#> 61: 5.952533        C
#> 62: 4.312156        C
#> 63: 4.087169        C
#> 64: 4.628592        C
#>           Cy CrecType
#> 
generateTestTbls(LETTERS[1:5]) # makes 5 tables with method CENSUS
#> $A
#>    Aid AnumSamp AnumTotal AselectMeth AstratumName Astratification       Ay
#> 1:   1        4         4      CENSUS            U               N 4.064187
#> 2:   2        4         4      CENSUS            U               N 3.928207
#> 3:   3        4         4      CENSUS            U               N 5.427932
#> 4:   4        4         4      CENSUS            U               N 3.068502
#>    ArecType
#> 1:        A
#> 2:        A
#> 3:        A
#> 4:        A
#> 
#> $B
#>     Bid BnumSamp BnumTotal BselectMeth BstratumName Bstratification Aid
#>  1:   1        4         4      CENSUS            U               N   1
#>  2:   2        4         4      CENSUS            U               N   1
#>  3:   3        4         4      CENSUS            U               N   1
#>  4:   4        4         4      CENSUS            U               N   1
#>  5:   5        4         4      CENSUS            U               N   2
#>  6:   6        4         4      CENSUS            U               N   2
#>  7:   7        4         4      CENSUS            U               N   2
#>  8:   8        4         4      CENSUS            U               N   2
#>  9:   9        4         4      CENSUS            U               N   3
#> 10:  10        4         4      CENSUS            U               N   3
#> 11:  11        4         4      CENSUS            U               N   3
#> 12:  12        4         4      CENSUS            U               N   3
#> 13:  13        4         4      CENSUS            U               N   4
#> 14:  14        4         4      CENSUS            U               N   4
#> 15:  15        4         4      CENSUS            U               N   4
#> 16:  16        4         4      CENSUS            U               N   4
#>           By BrecType
#>  1: 6.235257        B
#>  2: 5.752697        B
#>  3: 4.349598        B
#>  4: 4.326624        B
#>  5: 4.654513        B
#>  6: 5.422685        B
#>  7: 4.705715        B
#>  8: 5.489678        B
#>  9: 3.863877        B
#> 10: 4.286256        B
#> 11: 6.197262        B
#> 12: 3.213232        B
#> 13: 4.567375        B
#> 14: 3.774572        B
#> 15: 3.794069        B
#> 16: 4.274577        B
#> 
#> $C
#>     Cid CnumSamp CnumTotal CselectMeth CstratumName Cstratification Bid
#>  1:   1        4         4      CENSUS            U               N   1
#>  2:   2        4         4      CENSUS            U               N   1
#>  3:   3        4         4      CENSUS            U               N   1
#>  4:   4        4         4      CENSUS            U               N   1
#>  5:   5        4         4      CENSUS            U               N   2
#>  6:   6        4         4      CENSUS            U               N   2
#>  7:   7        4         4      CENSUS            U               N   2
#>  8:   8        4         4      CENSUS            U               N   2
#>  9:   9        4         4      CENSUS            U               N   3
#> 10:  10        4         4      CENSUS            U               N   3
#> 11:  11        4         4      CENSUS            U               N   3
#> 12:  12        4         4      CENSUS            U               N   3
#> 13:  13        4         4      CENSUS            U               N   4
#> 14:  14        4         4      CENSUS            U               N   4
#> 15:  15        4         4      CENSUS            U               N   4
#> 16:  16        4         4      CENSUS            U               N   4
#> 17:  17        4         4      CENSUS            U               N   5
#> 18:  18        4         4      CENSUS            U               N   5
#> 19:  19        4         4      CENSUS            U               N   5
#> 20:  20        4         4      CENSUS            U               N   5
#> 21:  21        4         4      CENSUS            U               N   6
#> 22:  22        4         4      CENSUS            U               N   6
#> 23:  23        4         4      CENSUS            U               N   6
#> 24:  24        4         4      CENSUS            U               N   6
#> 25:  25        4         4      CENSUS            U               N   7
#> 26:  26        4         4      CENSUS            U               N   7
#> 27:  27        4         4      CENSUS            U               N   7
#> 28:  28        4         4      CENSUS            U               N   7
#> 29:  29        4         4      CENSUS            U               N   8
#> 30:  30        4         4      CENSUS            U               N   8
#> 31:  31        4         4      CENSUS            U               N   8
#> 32:  32        4         4      CENSUS            U               N   8
#> 33:  33        4         4      CENSUS            U               N   9
#> 34:  34        4         4      CENSUS            U               N   9
#> 35:  35        4         4      CENSUS            U               N   9
#> 36:  36        4         4      CENSUS            U               N   9
#> 37:  37        4         4      CENSUS            U               N  10
#> 38:  38        4         4      CENSUS            U               N  10
#> 39:  39        4         4      CENSUS            U               N  10
#> 40:  40        4         4      CENSUS            U               N  10
#> 41:  41        4         4      CENSUS            U               N  11
#> 42:  42        4         4      CENSUS            U               N  11
#> 43:  43        4         4      CENSUS            U               N  11
#> 44:  44        4         4      CENSUS            U               N  11
#> 45:  45        4         4      CENSUS            U               N  12
#> 46:  46        4         4      CENSUS            U               N  12
#> 47:  47        4         4      CENSUS            U               N  12
#> 48:  48        4         4      CENSUS            U               N  12
#> 49:  49        4         4      CENSUS            U               N  13
#> 50:  50        4         4      CENSUS            U               N  13
#> 51:  51        4         4      CENSUS            U               N  13
#> 52:  52        4         4      CENSUS            U               N  13
#> 53:  53        4         4      CENSUS            U               N  14
#> 54:  54        4         4      CENSUS            U               N  14
#> 55:  55        4         4      CENSUS            U               N  14
#> 56:  56        4         4      CENSUS            U               N  14
#> 57:  57        4         4      CENSUS            U               N  15
#> 58:  58        4         4      CENSUS            U               N  15
#> 59:  59        4         4      CENSUS            U               N  15
#> 60:  60        4         4      CENSUS            U               N  15
#> 61:  61        4         4      CENSUS            U               N  16
#> 62:  62        4         4      CENSUS            U               N  16
#> 63:  63        4         4      CENSUS            U               N  16
#> 64:  64        4         4      CENSUS            U               N  16
#>     Cid CnumSamp CnumTotal CselectMeth CstratumName Cstratification Bid
#>           Cy CrecType
#>  1: 5.182337        C
#>  2: 5.420022        C
#>  3: 4.868424        C
#>  4: 6.105972        C
#>  5: 5.389740        C
#>  6: 5.714651        C
#>  7: 6.691967        C
#>  8: 5.169713        C
#>  9: 5.032683        C
#> 10: 4.077889        C
#> 11: 5.370198        C
#> 12: 4.458374        C
#> 13: 6.109346        C
#> 14: 5.153502        C
#> 15: 5.291783        C
#> 16: 5.095943        C
#> 17: 5.438530        C
#> 18: 5.842383        C
#> 19: 5.430552        C
#> 20: 5.699362        C
#> 21: 3.019886        C
#> 22: 4.440236        C
#> 23: 4.808955        C
#> 24: 6.039659        C
#> 25: 5.219161        C
#> 26: 6.067096        C
#> 27: 4.533845        C
#> 28: 5.800639        C
#> 29: 5.010433        C
#> 30: 5.457747        C
#> 31: 4.093050        C
#> 32: 3.516865        C
#> 33: 4.097791        C
#> 34: 3.601139        C
#> 35: 5.736757        C
#> 36: 4.718727        C
#> 37: 6.154610        C
#> 38: 3.083084        C
#> 39: 3.523824        C
#> 40: 4.946788        C
#> 41: 4.110649        C
#> 42: 5.741103        C
#> 43: 4.141768        C
#> 44: 5.346333        C
#> 45: 3.546911        C
#> 46: 4.536447        C
#> 47: 4.879987        C
#> 48: 3.672881        C
#> 49: 4.593583        C
#> 50: 4.924428        C
#> 51: 5.519728        C
#> 52: 5.535396        C
#> 53: 5.116807        C
#> 54: 5.454297        C
#> 55: 5.104514        C
#> 56: 4.044565        C
#> 57: 5.910741        C
#> 58: 4.691653        C
#> 59: 3.341418        C
#> 60: 4.487151        C
#> 61: 5.969622        C
#> 62: 5.688208        C
#> 63: 5.485796        C
#> 64: 4.579850        C
#>           Cy CrecType
#> 
#> $D
#>      Did DnumSamp DnumTotal DselectMeth DstratumName Dstratification Cid
#>   1:   1        4         4      CENSUS            U               N   1
#>   2:   2        4         4      CENSUS            U               N   1
#>   3:   3        4         4      CENSUS            U               N   1
#>   4:   4        4         4      CENSUS            U               N   1
#>   5:   5        4         4      CENSUS            U               N   2
#>  ---                                                                    
#> 252: 252        4         4      CENSUS            U               N  63
#> 253: 253        4         4      CENSUS            U               N  64
#> 254: 254        4         4      CENSUS            U               N  64
#> 255: 255        4         4      CENSUS            U               N  64
#> 256: 256        4         4      CENSUS            U               N  64
#>            Dy DrecType
#>   1: 4.725913        D
#>   2: 5.143399        D
#>   3: 3.990391        D
#>   4: 5.113538        D
#>   5: 4.921263        D
#>  ---                  
#> 252: 4.404846        D
#> 253: 4.166440        D
#> 254: 5.882933        D
#> 255: 4.657648        D
#> 256: 5.236679        D
#> 
#> $E
#>        Eid EnumSamp EnumTotal EselectMeth EstratumName Estratification Did
#>    1:    1        4         4      CENSUS            U               N   1
#>    2:    2        4         4      CENSUS            U               N   1
#>    3:    3        4         4      CENSUS            U               N   1
#>    4:    4        4         4      CENSUS            U               N   1
#>    5:    5        4         4      CENSUS            U               N   2
#>   ---                                                                     
#> 1020: 1020        4         4      CENSUS            U               N 255
#> 1021: 1021        4         4      CENSUS            U               N 256
#> 1022: 1022        4         4      CENSUS            U               N 256
#> 1023: 1023        4         4      CENSUS            U               N 256
#> 1024: 1024        4         4      CENSUS            U               N 256
#>             Ey ErecType
#>    1: 6.068461        E
#>    2: 3.474623        E
#>    3: 4.082617        E
#>    4: 4.283264        E
#>    5: 6.584557        E
#>   ---                  
#> 1020: 4.957682        E
#> 1021: 5.402889        E
#> 1022: 5.768723        E
#> 1023: 5.214824        E
#> 1024: 5.077314        E
#>